ocr-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
easyocr==1.7.1
Pillow==10.2.0
